# protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Robot Local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_v1_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pairing": {
      "post": {
        "summary": "Invite",
        "operationId": "invite_v1_pairing_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pairing/claim": {
      "post": {
        "summary": "Claim",
        "operationId": "claim_v1_pairing_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairClaim"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/heartbeat": {
      "post": {
        "summary": "Heartbeat",
        "operationId": "heartbeat_v1_heartbeat_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Heartbeat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FailureCounts": {
        "properties": {
          "turn": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Turn",
            "default": 0
          },
          "media": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Media",
            "default": 0
          },
          "download": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Download",
            "default": 0
          },
          "network": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Network",
            "default": 0
          },
          "progress-overflow": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Progress-Overflow",
            "default": 0
          },
          "progress-rejected": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Progress-Rejected",
            "default": 0
          },
          "progress-response": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Progress-Response",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FailureCounts"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Heartbeat": {
        "properties": {
          "status": {
            "type": "string",
            "maxLength": 40,
            "title": "Status",
            "default": "standby"
          },
          "appliedVersion": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Appliedversion",
            "default": 0
          },
          "camera": {
            "type": "boolean",
            "title": "Camera",
            "default": false
          },
          "microphone": {
            "type": "boolean",
            "title": "Microphone",
            "default": false
          },
          "reason": {
            "type": "string",
            "maxLength": 80,
            "title": "Reason",
            "default": ""
          },
          "serviceFailures": {
            "$ref": "#/components/schemas/FailureCounts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Heartbeat"
      },
...
      "PairClaim": {
        "properties": {
          "invite": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "Invite"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite",
          "name"
        ],
        "title": "PairClaim"
      },
...
      "Register": {
        "properties": {
          "invite": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "Invite"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite",
          "name"
        ],
        "title": "Register"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}